requirements.txt
google-generativeai
tiktoken 
chromadb
langchain
langchain-google-genai
flask
GitPython
python-dotenv
-e .